CreateTeamCommandBaseHandler
