+++ src/OctoshiftCLI.Tests/Octoshift/Commands/CreateTeamCommandBaseTests.cs
@@ -13,7 +13,7 @@ public class CreateTeamCommandBaseTests
     private readonly Mock<ITargetGithubApiFactory> _mockGithubApiFactory = new();
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-    private readonly CreateTeamCommandBase _command;
+    private readonly CreateTeamCommandBaseHandler _command;
 
     private const string GITHUB_ORG = "FooOrg";
     private const string TEAM_NAME = "foo-team";
@@ -24,7 +24,7 @@ public class CreateTeamCommandBaseTests
 
     public CreateTeamCommandBaseTests()
     {
-        _command = new CreateTeamCommandBase(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+        _command = new CreateTeamCommandBaseHandler(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
     }
 
     [Fact]
